+++ sockets_tuto/TrameAnalyzer.h
@@ -3,22 +3,19 @@
 
 class TrameAnalyzer
 {
-
 private:
 
 	std::string trame;
 	std::string date;
-	float temp;
-	float humidite;
-	float humiditeCorrigee;
+	double temp;
+	double humidite;
+	double humiditeCorrigee;
 
 public:
-
 	TrameAnalyzer(char trame[]);
-
 	std::string getDate();
-	float getTemp();
-	float getHumidite();
-	float gethumiditeCorrigee();
+	double getTemp();
+	double getHumidite();
+	double gethumiditeCorrigee();
 };
